CW-1149 Added Loader if user open new chat. Refactoring
