+++ src/pages/common/components/ChatComponent/components/ChatContent/ChatContent.module.scss
@@ -1,0 +1,33 @@
+@import "../../../../../../constants.scss";
+@import "../../../../../../styles/sizes";
+
+.messageList {
+  scrollbar-width: none;
+  padding: 0;
+  margin: 0;
+
+  & > li {
+    list-style: none;
+  }
+}
+
+.dateTitle {
+  padding: 0.0625rem 0.1875rem;
+  border-radius: 0.25rem;
+  background-color: $c-neutrals-100;
+  font-size: $xxsmall;
+  text-align: center;
+  color: $c-primary-300;
+  margin: 0 auto;
+  width: fit-content;
+  margin: 1.3125rem auto 1rem;
+}
+
+.newMessageLoaderWrapper {
+  width: 100%;
+  height: 6rem;
+}
+
+.veryFirstMessage {
+  margin-top: 5rem;
+}
