+++ src/pages/common/components/ChatComponent/components/ChatContent/index.ts
@@ -1,0 +1,1 @@
+export { default as ChatContent } from "./ChatContent";
