+++ src/pages/common/components/ChatComponent/index.scss
@@ -9,45 +9,10 @@
   position: relative;
   box-sizing: border-box;
 
-  .messages {
+  .loader-container {
+    display: flex;
     height: 100%;
-    overflow: auto;
-
-    &.empty {
-      display: flex;
-      justify-content: center;
-    }
-
-    .message-list {
-      scrollbar-width: none;
-      padding: 0;
-      margin: 0;
-
-      > li {
-        list-style: none;
-      }
-
-      .date-title {
-        padding: 0.0625rem 0.1875rem;
-        border-radius: 4px;
-        background-color: $c-neutrals-100;
-        font-size: 12px;
-        text-align: center;
-        color: $c-primary-300;
-        margin: 0 auto;
-        width: fit-content;
-        margin: 1.3125rem auto 1rem;
-      }
-    }
-
-    .new-message-loader-wrapper {
-      width: 100%;
-      height: 6rem;
-
-      &.very-first-message {
-        margin-top: 5rem;
-      }
-    }
+    align-items: center;
   }
 
   .bottom-chat-wrapper {
